Module integration, feature extension, CLI menu overhaul, Now an interactive, smarter tool
